Use fixed version of Ubuntu
Hi, there's a small issue in the Dockerfile. It's implicitly referencing `:latest` tag which have just been recently updated to Ubuntu 22.x (Jammy) and build is now failing because node v16 is not yet supported there.

This should get fixed soon but it's usually better idea not to reference floating tags like `latest` as this tends to break in time.

When set to `:focal` it builds correctly.